update example to 0.34
