+++ examples/FacebookTabsExample/package.json
@@ -7,7 +7,7 @@
   },
   "dependencies": {
     "react": "^15.1.0",
-    "react-native": "^0.28.0",
+    "react-native": "^0.34.1",
     "react-native-scrollable-tab-view": "../..",
     "react-native-vector-icons": "^2.0.3",
     "react-timer-mixin": "^0.13.3"
